+++ packages/react-components/src/index.tsx
@@ -1,3 +1,4 @@
+export { default as applyRefs } from './applyRefs';
 export { default as Button } from './Button';
 export { default as Calendar } from './Calendar';
 export { default as CardFooterButton } from './CardFooterButton';
@@ -17,6 +18,7 @@ export { default as Join } from './Join';
 export { default as JSONInput } from './JSONInput';
 export { default as Loader } from './Loader';
 export { default as Location } from './Location';
+export { default as MarkdownContent } from './MarkdownContent';
 export { default as Message } from './Message';
 export { default as MessagesProvider } from './MessagesProvider';
 export { default as Modal } from './Modal';
@@ -36,4 +38,5 @@ export { default as useMessages, MessagesContext } from './hooks/useMessages';
 export { default as useObjectURL } from './hooks/useObjectURL';
 export { default as useQuery } from './hooks/useQuery';
 export { default as useToggle } from './hooks/useToggle';
+export { default as useLocationString } from './hooks/useLocationString';
 export type { Toggle } from './hooks/useToggle';